+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -595,8 +595,8 @@ class GetOnePromptResponseBodyTypedDict(TypedDict):
     r"""The prompt’s name, meant to be displayable in the UI."""
     prompt_config: GetOnePromptPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
-    created_by_id: NotRequired[str]
-    updated_by_id: NotRequired[str]
+    created_by_id: NotRequired[Nullable[str]]
+    updated_by_id: NotRequired[Nullable[str]]
     description: NotRequired[Nullable[str]]
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
     metadata: NotRequired[GetOnePromptMetadataTypedDict]
@@ -623,9 +623,9 @@ class GetOnePromptResponseBody(BaseModel):
     prompt_config: GetOnePromptPromptConfig
     r"""A list of messages compatible with the openAI schema"""
 
-    created_by_id: Optional[str] = None
+    created_by_id: OptionalNullable[str] = UNSET
 
-    updated_by_id: Optional[str] = None
+    updated_by_id: OptionalNullable[str] = UNSET
 
     description: OptionalNullable[str] = UNSET
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
@@ -635,7 +635,7 @@ class GetOnePromptResponseBody(BaseModel):
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = ["created_by_id", "updated_by_id", "description", "metadata"]
-        nullable_fields = ["description"]
+        nullable_fields = ["created_by_id", "updated_by_id", "description"]
         null_default_fields = []
 
         serialized = handler(self)