+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -617,8 +617,8 @@ class GetPromptVersionResponseBodyTypedDict(TypedDict):
     prompt_config: GetPromptVersionPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     timestamp: str
-    created_by_id: NotRequired[str]
-    updated_by_id: NotRequired[str]
+    created_by_id: NotRequired[Nullable[str]]
+    updated_by_id: NotRequired[Nullable[str]]
     description: NotRequired[Nullable[str]]
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
     metadata: NotRequired[GetPromptVersionMetadataTypedDict]
@@ -634,9 +634,9 @@ class GetPromptVersionResponseBody(BaseModel):
 
     timestamp: str
 
-    created_by_id: Optional[str] = None
+    created_by_id: OptionalNullable[str] = UNSET
 
-    updated_by_id: Optional[str] = None
+    updated_by_id: OptionalNullable[str] = UNSET
 
     description: OptionalNullable[str] = UNSET
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
@@ -646,7 +646,7 @@ class GetPromptVersionResponseBody(BaseModel):
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = ["created_by_id", "updated_by_id", "description", "metadata"]
-        nullable_fields = ["description"]
+        nullable_fields = ["created_by_id", "updated_by_id", "description"]
         null_default_fields = []
 
         serialized = handler(self)