+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -79,5 +79,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-03T11:13:50.090Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T22:54:38.560Z")
     r"""The date and time the resource was last updated"""